Fix duplicate copy of moment.js in memory
Summary:
Ok, this was a hard issue to track down! The issue came up due to a change I made in Jan 2016 to always expose the `moment` object as a global (to ensure that exercises would continue to work). However, when I defined it I called the `factory()` function again, meaning that the copy of moment that you got when you called `require("moment")` was different from the one that you got when you called `window.moment`.

This became a problem for the locale plugins to moment as they didn't have `"expose_require": True,` set in `third_party_js.py` so they would look for moment at `window.moment` and extend it. Thus the copy of moment that we got back from `require("moment")` wouldn't have any non-English locale info associated with it. I fixed the problem my only assigning to the global if `module.exports` exists (which was the edge case we cared about anyway).

We could change the locale plugins to have `"expose_require": True,` set in `third_party_js.py` but the plugins all refer to moment by doing `require("../moment")` and this seems to confuse our build system. I'd really prefer to not be depending on exposed globals but at the very least this allows things to work as expected.

Asana task: https://app.asana.com/0/29899453652652/387484852510437

Test Plan:
Logged in as a coach I visited a class of mine and saw some assignments listed:
http://localhost:8080/coach/class/5584144679567360/assignments?lang=boxes

I confirmed that the date strings were translated to the boxes formatted:
{F782811}

(I'll be including the boxes locale plugin in a separate diff.)

Reviewers: csilvers, emily, kevinb

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D37977