Adding sr-cyrl.js back in moment-khansrc/locale so that our moment js mappings work as expected
This tweak is to add back sr-cyrl.js, which was renamed to sr.js because there wasnt a better way to map sr to sr-cyrl until now.  So the current expectation is that sr-cyrl.js which is the file that comes with stock momentjs, is the file we want to use rather then sr.js.  The mapping from sr.js to sr-cyrl.js is handled in a convertion call done on the server side in webapp.

Test Plan:
1. Start up your local server - `make serve`
2. Run the following query in `http://localhost:8080/devadmin/graphql`

```
query {
	locales{
    kaLocale
    momentLocale
    icuLocale
  }
}
```

Verify that momentLocale is `sr-cyrl`.