Removed remainder of the moment sr locale file renaming from previous work in D47686
Summary:
This diff renames sr-latin.js back to sr.js, which is what we get with stock moment js.  We have added logic in the server to properly do the sr-cyrl convertion for us, which makes these changes in moment no longer necessary.  The renaming was introduced in

https://github.com/Khan/moment/commit/258b8f07313e65e6831f7a2ff845b0c25a6385af
https://phabricator.khanacademy.org/D47686

Test Plan:
1. Start up your local server - `make serve`
2. Load up any page where you can access notifications, which can easily show sr-cyrl being used by default. Dont use dev webpack since this is for testing kake!
3. Verify that the notifications have Cyrillic in the time they show.  See attached screenshots.

Reviewers: csilvers, john

Subscribers: